[REF] VRange: move public things related to range to a new "range" utils
Public functions, types, constants and variables related to range were
spread between `Range.ts`, `VRange.ts` and `VDocument.ts` (`withRange`).
That quickly caused circular dependency issues when making VNodes
modular, which motivated this small refactoring.
