--- packages/core/src/Range.ts
@@ -1,8 +1,0 @@
-type Direction = 'ltr' | 'rtl';
-export interface Range {
-    startContainer: Node;
-    startOffset: number;
-    endContainer: Node;
-    endOffset: number;
-    direction: Direction;
-}
