+++ packages/plugin-devtools/src/components/TreeComponent.ts
@@ -2,7 +2,8 @@ import { utils } from '../../../../packages/utils/src/utils';
 import { OwlUIComponent } from '../../../owl-ui/src/OwlUIComponent';
 import { isRange } from '../../../../packages/utils/src/Predicates';
 import { VNode, VNodeType } from '../../../core/src/VNode';
-import { VRangeDescription, Direction } from '../../../core/src/VRange';
+import { VRangeDescription } from '../../../core/src/VRange';
+import { Direction } from '../../../utils/src/range';
 
 interface NodeProps {
     isRoot: boolean;
